add version metadata

+++ internals/shared.js
@@ -16,6 +16,11 @@ function verifyHash(folder, pubKey) {
     const encHash = Buffer.from(JSON.parse(fs.readFileSync(path.join(folder,'version.json'))).hash, 'base64');
     return getPubHash(encHash, pubKey) === require('./hash_folder.js')(folder, []).hash.toString('base64');
 }
+function verifyVersion(version, folder) {
+    const reqVer = JSON.parse(fs.readFileSync(path.join(folder,'version.json'))).clawffee_version;
+    return version != reqVer;
+}
+
 
 
 const pubKey = require('../internal.pub.txt')?.default || require('../internal.pub.txt');
@@ -62,6 +67,7 @@ function runUpdate() { return new Promise((resolve, reject) => {update_info.then
     function verifyDownload() {
         console.log(`finished inflating update at ${folderPath}`);
         if(!verifyHash(folderPath, pubKey)) return reject('Hash of downloaded folder is incorrect!!!');
+        if(!verifyVersion(globalThis.clawffeeInternals.launcher.meta.version, folderPath)) return reject('Clawffee executable outdated for the update, please download the newest Clawffee executable manually if you wish to update!')
         try {
             fs.rmSync('plugins/internal.bak', {force: true, recursive: true});
         } catch(e) {} // can silently fail
